Refs #008 Reorganized tree with demos
The directory tree with the demo apps is now ordered by progression in
the API usage. This makes it more clear what parts of the Avrtk API
are used by the demos.
