+++ demos/002-Blink101/src/main/c/BlinkWithFade.c
