+++ demos/008-AdcGauge101/Makefile
@@ -1,0 +1,24 @@
+###########################################################################
+#
+# Copyright (c) 2014-2016 Jorge Nunes, All Rights Reserved.
+#
+###########################################################################
+
+B_PROJ_ROOT = ../..
+
+include $(B_PROJ_ROOT)/make/avrtk.make
+
+
+B_PROJ = 008-AdcGauge101
+
+
+#
+# The main sources, to be compiled with AVR-GCC.
+#
+
+B_MAIN_C_SOURCES = \
+	AdcGauge101.c
+
+
+include $(B_PROJ_ROOT)/make/avrtk-demo.make 
+
