--- demos/BlinkManyVariable/Makefile
@@ -1,24 +1,0 @@
-###########################################################################
-#
-# Copyright (c) 2015 Jorge Nunes, All Rights Reserved.
-#
-###########################################################################
-
-B_PROJ_ROOT = ../..
-
-include $(B_PROJ_ROOT)/make/avrtk.make
-
-
-B_PROJ = BlinkManyVariable
-
-
-#
-# The main sources, to be compiled with AVR-GCC.
-#
-
-B_MAIN_C_SOURCES = \
-	BlinkManyVariable.c
-
-
-include $(B_PROJ_ROOT)/make/avrtk-demo.make 
-
