--- demos/BlinkVariable101/Makefile
@@ -1,24 +1,0 @@
-###########################################################################
-#
-# Copyright (c) 2014 Jorge Nunes, All Rights Reserved.
-#
-###########################################################################
-
-B_PROJ_ROOT = ../..
-
-include $(B_PROJ_ROOT)/make/avrtk.make
-
-
-B_PROJ = BlinkVariable101
-
-
-#
-# The main sources, to be compiled with AVR-GCC.
-#
-
-B_MAIN_C_SOURCES = \
-	BlinkVariable101.c
-
-
-include $(B_PROJ_ROOT)/make/avrtk-demo.make 
-
